+++ server/models/day_model.js
@@ -2,13 +2,15 @@ const mongoose = require('mongoose')
 const aggregatePaginate = require('mongoose-aggregate-paginate-v2')
 require('dotenv').config()
 
+const Schema = mongoose.Schema
+
 // subdocument to specify post and
-const content_Schema = new mongoose.Schema({
+const content_Schema = new Schema({
   text: String,
   status: Number,
 })
 
-const day_PostSchema = mongoose.Schema({
+const day_PostSchema = Schema({
   content: [content_Schema],
 
   date: {